requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-socketio==5.10.0
python-multipart==0.0.6
jinja2==3.1.2
groq==0.4.1
python-dotenv==1.0.0
transformers>=4.30.0
torch>=1.13.0
soundfile>=0.12.0
numpy>=1.21.0
speechbrain>=0.5.0
datasets>=2.0.0